% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaliMRCT.R
\name{CaliMRCT}
\alias{CaliMRCT}
\title{CaliMRCT}
\usage{
CaliMRCT(R, X, A, Y, Nboot = 300, seed = 2022)
}
\arguments{
\item{R}{Region indicator}

\item{X}{Matrix of m baseline covariates (after standardization)}

\item{A}{Treatment indicator}

\item{Y}{Observed outcome}

\item{Nboot}{Number of bootstrap samples used (default is 300)}

\item{seed}{Pseudorandom number for reproducible results (default is 2022)}
}
\value{
A list containing
\itemize{
\item naive: a matrix containing estimated ATEs with bootstrap SEs and confidence intervals by Naive method
\item ipw: a matrix containing estimated ATEs with bootstrap SEs and confidence intervals by IPW method
\item cw: a matrix containing estimated ATEs with bootstrap SEs and confidence intervals by CW method
\item cwe: a matrix containing estimated ATEs with bootstrap SEs and confidence intervals by CWE method
\item tab: a table including results by all four methods
}
}
\description{
Calibration Weighting Estimation of Treatment Effect Heterogeneity for MRCTs
}
\details{
CaliMRCT is a function to assess treatment effect heterogeneity for multi-regional randomized clinical trials.
}
\examples{
CaliMRCT(R = simMRCT$R, X = cbind(simMRCT$X1,simMRCT$X2), A = simMRCT$A, Y = simMRCT$Y, Nboot = 300, seed = 2022)
}
\author{
Haotian Zhuang, Xiaofei Wang \href{mailto:haotian.zhuang@duke.edu}{haotian.zhuang@duke.edu}
}
