% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMRCT.R
\docType{data}
\name{simMRCT}
\alias{simMRCT}
\title{Multi-regional randomized clinical trials (MRCTs)}
\format{
A data frame with 800 subjects and 5 variables.
\itemize{
\item R: region
\item X1: baseline covariate 1
\item X2: baseline covariate 2
\item A: treatment received
\item Y: observed outcome
}
}
\usage{
simMRCT
}
\description{
A simulated dataset contains 800 rows and 5 columns.
}
\keyword{datasets}
